+++ src/CreatioHelper.Shared/Utils/Constants.cs
@@ -4,5 +4,7 @@ public static class Constants
 {
     public static readonly Version MinimumVersionForDeletePackages = new Version(8, 0, 4);
 
+    public static readonly Version MinimumVersionForFastCompile = new Version(8, 0, 10);
+
     public static readonly Version MinimumVersionForRestoreConfiguration = new Version(8, 0, 2);
 }